+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* If Iguana Tweaks Reborn is installed, Breaching creepers now have bigger explosions
+
 ## 2.4.1
 * Fixed some mobs losing the ability to attack, like phantoms and wolves
 * Fixed Darkness range multiplier not being applied